Update folders refactor and change imports
