+++ lib/app/pages/home_page.dart
@@ -1,9 +1,9 @@
 import 'package:arqueo_ahsc/app/helpers/build_error_snack_bar.dart';
 import 'package:arqueo_ahsc/app/providers/day_cash_counts_provider.dart';
-import 'package:arqueo_ahsc/app/widgets/ahsc_logo.dart';
+import 'package:arqueo_ahsc/app/widgets/public/ahsc_logo.dart';
 import 'package:arqueo_ahsc/app/widgets/buttons/cancel_button.dart';
-import 'package:arqueo_ahsc/app/widgets/custom_drawer.dart';
-import 'package:arqueo_ahsc/app/widgets/day_cash_counts_list.dart';
+import 'package:arqueo_ahsc/app/widgets/drawer/custom_drawer.dart';
+import 'package:arqueo_ahsc/app/widgets/dayCashCounts/day_cash_counts_list.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
