+++ lib/app/widgets/public/ahsc_logo.dart
